11b: Solve 11b

+++ src/main/kotlin/aoc11/model/Operand.kt
@@ -2,19 +2,19 @@ package aoc11.model
 
 sealed interface Operand {
     object Old : Operand {
-        override fun invoke(old: Int) = old
+        override fun invoke(old: Long) = old
     }
 
-    data class Number(val number: Int) : Operand {
-        override fun invoke(old: Int) = number
+    data class Number(val number: Long) : Operand {
+        override fun invoke(old: Long) = number
     }
 
-    operator fun invoke(old: Int): Int
+    operator fun invoke(old: Long): Long
 
     companion object {
         fun of(operand: String) = when (operand) {
             "old" -> Old
-            else -> Number(operand.toInt())
+            else -> Number(operand.toLong())
         }
     }
 }
