+++ src/main/kotlin/aoc11/model/Operation.kt
@@ -2,14 +2,14 @@ package aoc11.model
 
 sealed class Operation(protected open val operands: Pair<Operand, Operand>) {
     data class Plus(override val operands: Pair<Operand, Operand>) : Operation(operands) {
-        override fun invoke(old: Int) = operands.first(old) + operands.second(old)
+        override fun invoke(old: Long) = operands.first(old) + operands.second(old)
     }
 
     data class Times(override val operands: Pair<Operand, Operand>) : Operation(operands) {
-        override fun invoke(old: Int) = operands.first(old) * operands.second(old)
+        override fun invoke(old: Long) = operands.first(old) * operands.second(old)
     }
 
-    abstract operator fun invoke(old: Int): Int
+    abstract operator fun invoke(old: Long): Long
 
     companion object {
         fun of(firstOperand: String, operator: String, secondOperand: String) = when (operator) {
